Add GitHub inspired file finder to search for and open any file
The file finder has been added to a new UI module since it can
potentially be reused for the tree and blame views. It is hooked into
the open_blob_view method, and will be launch if no file has been
selected or if the users presses the key to open the blob view when
already in the blob view.

In its current form it is not well integrated with the rest of the view
code, since the file finder creates a separate window that is overlayed
the current view. If possible, code for maintaining the current position
and offset should be shared between this and the other view code.

Fixes #342