Fix testframework
At least in my shell(s), "cd foo" echoes the new cwd to stdout so that
"cd foo && pwd" echoes the new cwd twice, which lets the test framework
fail completely (due to the wrong $source_dir and everything that
depends on it).

Make that construct "cd foo >/dev/null && pwd" to be super safe, rather
than simplifying to "cd foo" and relying on the behaviour of "cd".