Do not exit when the main view is reloaded due to option toggling
Option toggling can cause the main view to be reloaded before any lines
have been read, which triggered a code path which exits tig if the main
view is empty. Pass in a flag to the view_ops read method noting whether
the current load is being forcefully stopped.

Fixes #470