Use pkg-config to detect ncurses
Update ax_with_curses.m4 to latest version and use pkg-config as primary
detection mechanism for ncurses. Latest version of ax_with_curses.m4
supports pkg-config. This is important as libcurses is split into
libcurses.so and libtinfo.so on many distros and linking against
libcurses.so is not sufficiant. Pkg-config allows proper detection
of necessary libs.

Signed-off-by: Justin Lecher <jlec@gentoo.org>