Trying to fix graphs where the same commit is hit by a merge and a new branch, but the graph gives the impression the commit is not merged.
The old case rendered this result. Not that the bottom branch looks
like it terminates without being merged:

 ● Update NEWS with post 2.1.1 changes
 ●─╮ Merge pull request #457 from vivien/text-variable
 │ ● add a %(text) variable
 ●─╯ Merge pull request #439 from peff/fix-blame-args
 │ ● blame: allow user to specify rev arguments to blame
 ●─╯ Update OSX make config to find brew installed ncurses
 ● Remove unneeded calls to {def,reset}_prog_mode

With fix I get this:

 ● Update NEWS with post 2.1.1 changes
 ●─╮ Merge pull request #457 from vivien/text-variable
 │ ● add a %(text) variable
 ●─┤ Merge pull request #439 from peff/fix-blame-args
 │ ● blame: allow user to specify rev arguments to blame
 ●─╯ Update OSX make config to find brew installed ncurses
 ● Remove unneeded calls to {def,reset}_prog_mode

Related to https://github.com/jonas/tig/issues/419

My current solution is not nice: it depends on the order of evaluation in the
graph_symbol_to_XXX() functions so that the new case of graph_symbol_merge_branch()
is hit before graph_symbol_turn_left() or graph_symbol_merge() do.

The graph-generating code is complex, and I have not spent enough time in it to
know for sure my fix does not mess up other situations.