Only accept "-C" before a subcommand (#911)
Since commit ed36f75 (Add -C option to specify the working directory,
2019-03-21), we allow:

  tig -C/path/to/repo

to change into the repo directory before running any git commands. But
because of the way tig's command-line parser works, this eats any "-C"
we meant to pass to a git command, like:

  tig blame -C foo.c

which used to enable copy-detection in the blame. Let's require that -C
come before any sub-command. That will continue to allow:

  tig -C/path/to/repo

with no subcommand. It also means that using "-C" with a subcommand
looks like this:

  tig -C/path/to/repo blame foo.c

which matches git's global -C option to change directories. And of
course it allows:

  tig blame -C foo.c

to work as it used to prior to ed36f75, enabling copy detection.