Improve graph rendering in the default graphic mode
Using the same symbol for ordinary and boundary commits is misleading.
Keep 'o' for boundary commits only, it's similar to other graphic modes.

Use ACS_BULLET for ordinary commits, ACS_DIAMOND for merges and '@' for
the initial commit, which is similar to the symbols used in the UTF-8
graphic mode.

Update tests for the new graph symbols. ACS_BULLET corresponds to '~' in
the alternative charset.