Document how to use pipes and redirection with 'bind'
When my bind command using a pipe did not work, I grepped through the
docs for 'pipe' and 'redirection' with no results. Only after I found
out in the issues tracker that it was intentional, then I noticed the
solution in the examples, too.

Add a paragraph to make it more explicit.

Signed-off-by: Wolfram Sang <wsa@kernel.org>
Signed-off-by: Jonas Fonseca <jonas.fonseca@gmail.com>