Expose %(lineno_old), the line number before a diff was applied
Popular Git hostings sites like GitHub and GitLab show two columns with line
numbers next to diffs. The left one is the line number in the "old" version
of a file, and the right one is the line number in the "new" file - the file
after the diff was applied.

I use Tig to write review comments on GitLab.  To comment on specific lines
of a diff, the GitLab API requires both old and new line numbers.
Tig's %(lineno) provides the new line number.  This patch additionally
exposes the old line number as %(lineno_old).

When browsing a diff that has not yet been applied to the working tree,
this line number can be used to go to the selected line.  Naturally, this
is only precise for "deleted" and "context" lines, that are present in the
working tree.