Avoid depending on wcwidth in test by removing Variation Selector (#1065)
test/main/emoji-test fails when using a recent version of glibc, where
wcwidth(🐛) equals 2 (used to be 1 which is less correct).

This test used to include 🐛 and other emoji followed by the 
["Variation Selector"](https://unicode-table.com/en/FE0F/).

With the old wcwidth, Tig would render both characters, one cell each.
With the new wcwidth, Tig only renders the first one twice, in two cells.

As a result, the output of :save-display varies across systems.  Solving this
is not really in scope for Tig, since this is the domain of wcwidth and
Ncurses. Hence, remove the variation selectors. As far as I can tell they
are useless here anyway because the characters render the same way after
removing them.