Fix blame when opened from an initial diff view (#1135)
When running "tig show 529182c -- README.adoc", blames were started
on HEAD instead of 529182c.  Fix this by reading the commit line of
the displayed diff.  This also works when multiple commits are shown
("tig show --all").

Using %(commit) instead of parsing the commit line did unfortunately
not work because, because %(commit) can contain the wrong SHA for when
returning from a blame view to a stash view, see test/blame/stash-test.

Sadly, stashes don't contain commit lines, so we need to find out
their SHAs in another way. Once we fix %(commit), we can use it,
but for now this commit adds a workaround to use the old
"view->vid", which seems to be set in the cases where display
the output from "git stash show".

Also, when blame is requested on a context line, use the current
commit rather than its parent.  This does not change behavior since
a context line can never be blamed on the current commit, but this
is consistent with the stage view, which only uses HEAD for deletions
and the index for others.

[ja: add test and commit message]

Co-authored-by: Thomas Koutcher <thomas.koutcher@online.fr>