Fix null dereferences on unset %(fileargs)
Commit 9b262f6b (Fix null dereferences on unset state variable, 2021-07-24)
fixed crashes on ":echo %(cmdlineargs)" but not on ":echo %(fileargs)".
(which is embarassing because the latter is even mentioned in the commit
message, but not in the test).

The problem is that it fixed only format_append_argv(), which was
used for %(cmdlineargs), but not for %(fileargs).  For %(fileargs),
argv_format() calls argv_append_array() (to append arguments verbatim).

When Tig is launched without arguments, opt_file_args is NULL, so, on
":echo %(fileargs)", argv_format() does not append anything to the
argv-style output array.  argv_format() reports success, but leaves
the output at NULL.  Callers like concat_argv() however expect an
argv-style array.

Handle this case in argv_format().  Make sure that a successful
argv_format() always return an argv-style array.  This should make
things simpler going forward.

This allows/requires a couple of other changes:
1. src/argv.c::format_append_argv(): we can revert this change from
   9b262f6b because this fix is more general.
2. src/prompt.c::run_prompt_command(): to handle "!%(cmdlineargs)"
   gracefully, we add a check that fails if next->argv[0] is NULL.
   Without this, Tig would hang (but not consume CPU).  I don't know
   why this happens but I doubt it's worth investigating, since
   an empty argument array means the command is invalid anyway.
   This causes the change to the bang-cmdlineargs-doesnt-crash test.
3. src/prompt.c::run_prompt_command(): since argv_format() now returns
   valid argv-style arrays on empty input, we no longer need the
   check if the first argument to echo is empty. That check was
   also misleading: it led me to think that the subsequent call to
   argv_format() would *only* format argv[1], when in reality it will
   format argv[1], argv[2], etc., until the NULL terminator.
4. src/prompt.c::exec_run_request(): we need to adjust the check
   for the case where argv_format() succeeded but returns an array
   of length zero - now we actually get an array instead of NULL.

An alternative fix would be to just declare empty results by
argv_format() as failure, but that feels wrong for ":echo"