Allow to toggle revision filtering
"tig -- <pathspec>" shows only commits that touch files matching
<pathspec>.  The command ":toggle file-filter" (bound to %) can be
used to toggle this filtering dynamically.

"tig <revisions range>" shows only commits in the given range.
This patch adds ":toggle rev-filter" (bound to ^) to quickly toggle
this filter.

There is already ":toggle rev-args" which is meant to disable rev
filtering but that does not work because it fails to redraw the
view. Also it cannot re-enable filtering.
Also ":set rev-args = HEAD" works but doesn't redraw the view either.
We should fix that (should be easy but I didn't check yet), and make
":toggle rev-args" reversible, then both the rev-filter and file-filter
toggles will be obsolete.

This feature is particularly useful with a new feature in Tig 2.5.2:
when pressing m in the blame view, Tig will show the blamed commit
centered in the main view, making it easy to explore adjacent
commits. Note that if there are many commits between the blamed commit
and HEAD, you'll need to wait a bit after pressing m.