Use the full width for diffstat in the stage view
The diff-files and diff-index commands used in the stage view need to
have the screen width passed explicitly with --stat argument to generate
an adjusted diffstat, as they ignore the COLUMNS environment variable.

The view arguments are not reevaluated upon refresh, so insert
`--stat=<width>` just before calling io_exec(), whenever `--stat` or
`--patch-with-stat` is present.