Automatically show next diff in the status view
Move view_request() to view.c where it belongs and, when changing a file
status in a split stage view spawned from the status view, use it to
simulate pressing the Enter key in the parent view to display the next
diff.

Note: the behaviour is similar to that of the Up and Down keys, the
action has to be performed from the child view.

Closes #413, closes #469