Remove redundant call to ncurses set_tabsize()
ncurses defines a global variable called TABSIZE which is the number
of columns occupied by a tab character. We set this option to the
value of the tab-size option.  I believe we never rely on TABSIZE
because we always expand tabs to the appropriate number of spaces
before passing them to ncurses.

The following patches will potentially give each screen its own
tab-size, which is not possible with the global TABSIZE, so we should
not be using it anyway.