Support editing from the pager view (#1243)
When working with pull requests, before merging to `master` it is nice to get the total diff of all of the commits of the PR, as in the `Files changed` tab on github. This is possible with `tig`, i.e. `git diff master | tig` or `:!git diff master`, but currently, this view does not support editing.

Make editing in the pager view possible.