Initialise struct timezone
On musl libc gettimeofday (which is used by time_now) does not
populate the timezone struct passed to it as POSIX says:

> If tzp is not a null pointer, the behavior is unspecified.

tz_minuteswest is later multiplied by 60 which can overflow. When tig
is compiled with integer overflow hardening (as is done on Chimera
Linux) via the clang option -fsanitize=signed-integer-overflow this can
result in tig crashing due to the overflow.