Don't return early for non-line cases
The prior fix was pretty blunt: return early if line is null.
However, line isn't really used unless specific conditions are met.
Instead of returning early, add a check for the value of line
where it's first used.

This also allows reducing the scope of the corresponding variables
from function to a specific block.