Blame: improve error messages for the `,` operation
The message is now:

   Commit f4657f00 has no ancestors modifying the seleted line

instead of  

   The selected commit has no parents

This addresses a portion of https://github.com/jonas/tig/issues/1315.

Apart from being a bit clearer, the user will be able to learn what `,`
is meant to do from it. In particular, they should notice that commit
`f4657f00` is the commit on the selected *line*, not the commit that is
currently displayed in general.

Trying to press `,` a few times is a natural thing to try to do after
reading the help while trying to figure out how blame works in tig.