Show committer date by default in the date column
Although the author name is generally of more interest than the
committer name, the committer date is more significant as it shows
when the change was actually added to the commit history. Make it
the default for the date column.

The author date can be restored on a per-view basis using the
`use-author` option of the `date` column in the view settings
of ~/.tigrc. See tigrc(5).

Note: to have both author date and committer date shown in the log
view, add `--pretty=fuller` to `log-options`.

Closes #294