POC: Open diff view from blame at the correct line
When pressing <Enter> (AKA `:enter`) on a line in the blame view,
previously the diff view would open with the cursor either on the first
line or on whatever line was last selected in that diff view (if it was
previously opened). This commit makes the diff view open at the line
corresponding to the line that was selected in the blame view.

One can still press <d> (AKA `:view-diff`) to open the diff view at
either the first line or the previously selected line.

The implementation is currently proof-of-concept quality: it works,
is tested, but does not fit well with the rest of the code. I'd need
help to make the memory management more robust and to fit better with
tig's design.

Also, I currently re-run the `git diff` command every time the diff view
is opened from blame, since I couldn't figure out an easier way to have
the view load without using the line it was previously opened at.